% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.EBrMAP}
\alias{plot.EBrMAP}
\title{Plot an EBrMAP Object}
\usage{
\method{plot}{EBrMAP}(x, ...)
}
\arguments{
\item{x}{"EBrMAP" class object returned by `EB_rMAP` function}

\item{...}{Ignored}
}
\value{
A ggplot2 object which can be further customized
}
\description{
Plot EB-rMAP weight against various observed response in the current trial
}
