% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wEB.R
\name{wEB}
\alias{wEB}
\title{Obtain the EB-rMAP Weight}
\usage{
wEB(obj, y_obs)
}
\arguments{
\item{obj}{"EBrMAP" class object returned by `EB_rMAP` function}

\item{y_obs}{Observed response in the current trial}
}
\value{
A scalar of EB-rMAP weight
}
\description{
Obtain the EB-rMAP weight corresponding to an observed response in current trial
}
\details{
`y_obs` for binomial and Poisson outcomes must be an integer.

In the case of normal outcome, `y_obs` is matched to the closest break in `y_range` argument in `EB_rMAP` function, whose corresponding EB-rMAP weight is returned. Rounding issues may cause error occasionally. In this case, jittering `y_obs` might help. For example, try `wEB(obj, 30.001)` if `wEB(obj, 30)` fails.
}
